working on validation on person and user

+++ src/main/java/com/easy_split/demo/repositories/PersonRepository.java
@@ -4,6 +4,6 @@ import com.easy_split.demo.entities.Person;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 
 }
